+++ deploy/ONNX/export_onnx.py
@@ -13,6 +13,7 @@ if str(ROOT) not in sys.path:
     sys.path.append(str(ROOT))
 
 from yolov6.models.yolo import *
+from yolov6.models.effidehead import EffiDeHead
 from yolov6.layers.common import *
 from yolov6.utils.events import LOGGER
 from yolov6.utils.checkpoint import load_checkpoint
@@ -52,7 +53,7 @@ if __name__ == '__main__':
         if isinstance(m, Conv):  # assign export-friendly activations
             if isinstance(m.act, nn.SiLU):
                 m.act = SiLU()
-        elif isinstance(m, Detect):
+        elif isinstance(m, EffiDeHead):
             m.inplace = args.inplace
        
     y = model(img)  # dry run
