+++ yolov6/models/yolo.py
@@ -1,117 +1,28 @@
 #!/usr/bin/env python3
 # -*- coding:utf-8 -*-
 import math
-import torch
 import torch.nn as nn
 from yolov6.layers.common import *
 from yolov6.utils.torch_utils import initialize_weights
 from yolov6.models.efficientrep import EfficientRep
 from yolov6.models.reppan import RepPANNeck
-
-
-class Detect(nn.Module):
-    stride = [8, 16, 32]  # strides computed during build
-
-    def __init__(self, nc=80, anchors=1, nl=3, inplace=True):  # detection layer
-        super().__init__()
-        self.nc = nc  # number of classes
-        self.no = nc + 5  # number of outputs per anchor
-        self.nl = nl  # number of detection layers
-        if isinstance(anchors, (list, tuple)):
-            self.na = len(anchors[0]) // 2
-        else:
-            self.na = anchors
-        self.anchors = anchors
-        self.grid = [torch.zeros(1)] * self.nl
-        self.prior_prob = 1e-2
-        self.inplace = inplace
-        
-        # Init decouple head
-        self.cls_convs = nn.ModuleList()
-        self.reg_convs = nn.ModuleList()
-        self.cls_preds = nn.ModuleList()
-        self.reg_preds = nn.ModuleList()
-        self.obj_preds = nn.ModuleList()
-        self.stems = nn.ModuleList()
-
-    def initialize_biases(self):
-        for conv in self.cls_preds:
-            b = conv.bias.view(self.na, -1)
-            b.data.fill_(-math.log((1 - self.prior_prob) / self.prior_prob))
-            conv.bias = torch.nn.Parameter(b.view(-1), requires_grad=True)
-
-        for conv in self.obj_preds:
-            b = conv.bias.view(self.na, -1)
-            b.data.fill_(-math.log((1 - self.prior_prob) / self.prior_prob))
-            conv.bias = torch.nn.Parameter(b.view(-1), requires_grad=True)
-               
-    def forward(self, x):
-        z = []
-        for i in range(self.nl):
-           
-            x[i] = self.stems[i](x[i])
-            cls_x = x[i]
-            reg_x = x[i]
-
-            cls_feat = self.cls_convs[i](cls_x)
-            cls_output = self.cls_preds[i](cls_feat)
-
-            reg_feat = self.reg_convs[i](reg_x)
-            reg_output = self.reg_preds[i](reg_feat)
-            obj_output = self.obj_preds[i](reg_feat)
-            
-            if self.training:
-                x[i] = torch.cat([reg_output, obj_output, cls_output], 1)
-                bs, _, ny, nx = x[i].shape
-                x[i] = x[i].view(bs, self.na, self.no, ny, nx).permute(0, 1, 3, 4, 2).contiguous()
-
-            else:
-                y = torch.cat([reg_output, obj_output.sigmoid(), cls_output.sigmoid()], 1)
-                bs, _, ny, nx = y.shape
-                y = y.view(bs, self.na, self.no, ny, nx).permute(0, 1, 3, 4, 2).contiguous()
-
-                if self.grid[i].shape[2:4] != y.shape[2:4]:
-                    d = self.stride.device
-                    yv, xv = torch.meshgrid([torch.arange(ny).to(d), torch.arange(nx).to(d)])
-                    self.grid[i] = torch.stack((xv, yv), 2).view(1, self.na, ny, nx, 2).float()
-
-                if self.inplace:
-                    y[..., 0:2] = (y[..., 0:2] + self.grid[i]) * self.stride[i]  # xy
-                    y[..., 2:4] = torch.exp(y[..., 2:4]) * self.stride[i] # wh
-                else:  
-                    xy = (y[..., 0:2] + self.grid[i]) * self.stride[i]  # xy
-                    wh = torch.exp(y[..., 2:4]) * self.stride[i]  # wh
-                    y = torch.cat((xy, wh, y[..., 4:]), -1)
-                z.append(y.view(bs, -1, self.no))    
-
-        return x if self.training else torch.cat(z, 1)
+from yolov6.models.effidehead import EffiDeHead, build_effidehead_layer
 
 
 class Model(nn.Module):
-    def __init__(self, config, ch=3, nc=None, anchors=None):  # model, input channels, number of classes
+    def __init__(self, config, channels=3, num_classes=None, anchors=None):  # model, input channels, number of classes
         super().__init__()
         # Build network
-        model, self.backbone, self.neck = build_network(config, ch, nc)
-        bi = config.model.head.begin_indices
-        nl = config.model.head.num_layers
-        oi_head = config.model.head.out_indices
+        num_layers = config.model.head.num_layers
+        self.backbone, self.neck, self.head = build_network(config, channels, num_classes, anchors, num_layers)
 
-        # Build Detect head
-        m = Detect(nc, anchors, nl)
-        m.stride = torch.tensor(m.stride)
-        self.stride = m.stride
-        m.i = bi
-        m.f = oi_head
-        for i in range(nl):
-            idx = i*6
-            m.stems.append(model[idx])
-            m.cls_convs.append(model[idx+1])
-            m.reg_convs.append(model[idx+2])
-            m.cls_preds.append(model[idx+3])
-            m.reg_preds.append(model[idx+4])
-            m.obj_preds.append(model[idx+5])
-        self.detect = m
-        self.detect.initialize_biases()
+        # Init Detect head
+        begin_indices = config.model.head.begin_indices
+        out_indices_head = config.model.head.out_indices
+        self.stride = self.head.stride
+        self.head.i = begin_indices
+        self.head.f = out_indices_head
+        self.head.initialize_biases()
 
         # Init weights
         initialize_weights(self)
@@ -119,13 +30,13 @@ class Model(nn.Module):
     def forward(self, x):
         x = self.backbone(x)
         x = self.neck(x)
-        x = self.detect(x)
+        x = self.head(x)
         return x
 
     def _apply(self, fn):
         self = super()._apply(fn)
-        self.detect.stride = fn(self.detect.stride)
-        self.detect.grid = list(map(fn, self.detect.grid))
+        self.head.stride = fn(self.head.stride)
+        self.head.grid = list(map(fn, self.head.grid))
         return self
 
 
@@ -134,7 +45,7 @@ def make_divisible(x, divisor):
     return math.ceil(x / divisor) * divisor
 
 
-def build_network(config, ic, nc):
+def build_network(config, channels, num_classes, anchors, num_layers):
     depth_mul = config.model.depth_multiple
     width_mul = config.model.width_multiple
     num_repeat_backbone = config.model.backbone.num_repeats
@@ -146,7 +57,7 @@ def build_network(config, ic, nc):
     channels_list = [make_divisible(i * width_mul, 8) for i in (channels_list_backbone + channels_list_neck)]
 
     backbone = EfficientRep(
-        in_channels=ic,
+        in_channels=channels,
         channels_list=channels_list,
         num_repeats=num_repeat
     )
@@ -156,135 +67,13 @@ def build_network(config, ic, nc):
         num_repeats=num_repeat
     )
 
-    effidehead_layers = build_effidehead_layer(channels_list, num_anchors, nc)
+    head_layers = build_effidehead_layer(channels_list, num_anchors, num_classes)
 
-    return effidehead_layers, backbone, neck
+    head = EffiDeHead(num_classes, anchors, num_layers, head_layers=head_layers)
 
-
-def build_effidehead_layer(channels_list, num_anchors, nc):
-    head_layers = nn.Sequential(
-        # stem0
-        Conv(
-            in_channels=channels_list[6],
-            out_channels=channels_list[6],
-            kernel_size=1,
-            stride=1
-        ),
-        # cls_conv0
-        Conv(
-            in_channels=channels_list[6],
-            out_channels=channels_list[6],
-            kernel_size=3,
-            stride=1
-        ),
-        # reg_conv0
-        Conv(
-            in_channels=channels_list[6],
-            out_channels=channels_list[6],
-            kernel_size=3,
-            stride=1
-        ),
-        # cls_pred0
-        nn.Conv2d(
-            in_channels=channels_list[6],
-            out_channels=nc * num_anchors,
-            kernel_size=1
-        ),
-        # reg_pred0
-        nn.Conv2d(
-            in_channels=channels_list[6],
-            out_channels=4 * num_anchors,
-            kernel_size=1
-        ),
-        # obj_pred0
-        nn.Conv2d(
-            in_channels=channels_list[6],
-            out_channels=1 * num_anchors,
-            kernel_size=1
-        ),
-        # stem1
-        Conv(
-            in_channels=channels_list[8],
-            out_channels=channels_list[8],
-            kernel_size=1,
-            stride=1
-        ),
-        # cls_conv1
-        Conv(
-            in_channels=channels_list[8],
-            out_channels=channels_list[8],
-            kernel_size=3,
-            stride=1
-        ),
-        # reg_conv1
-        Conv(
-            in_channels=channels_list[8],
-            out_channels=channels_list[8],
-            kernel_size=3,
-            stride=1
-        ),
-        # cls_pred1
-        nn.Conv2d(
-            in_channels=channels_list[8],
-            out_channels=nc * num_anchors,
-            kernel_size=1
-        ),
-        # reg_pred1
-        nn.Conv2d(
-            in_channels=channels_list[8],
-            out_channels=4 * num_anchors,
-            kernel_size=1
-        ),
-        # obj_pred1
-        nn.Conv2d(
-            in_channels=channels_list[8],
-            out_channels=1 * num_anchors,
-            kernel_size=1
-        ),
-        # stem2
-        Conv(
-            in_channels=channels_list[10],
-            out_channels=channels_list[10],
-            kernel_size=1,
-            stride=1
-        ),
-        # cls_conv2
-        Conv(
-            in_channels=channels_list[10],
-            out_channels=channels_list[10],
-            kernel_size=3,
-            stride=1
-        ),
-        # reg_conv2
-        Conv(
-            in_channels=channels_list[10],
-            out_channels=channels_list[10],
-            kernel_size=3,
-            stride=1
-        ),
-        # cls_pred2
-        nn.Conv2d(
-            in_channels=channels_list[10],
-            out_channels=nc * num_anchors,
-            kernel_size=1
-        ),
-        # reg_pred2
-        nn.Conv2d(
-            in_channels=channels_list[10],
-            out_channels=4 * num_anchors,
-            kernel_size=1
-        ),
-        # obj_pred2
-        nn.Conv2d(
-            in_channels=channels_list[10],
-            out_channels=1 * num_anchors,
-            kernel_size=1
-        )
-    )
-
-    return head_layers
+    return backbone, neck, head
 
 
 def build_model(cfg, num_classes, device):
-    model = Model(cfg, ch=3, nc=num_classes, anchors=cfg.model.head.anchors)
-    return model.to(device)
+    model = Model(cfg, channels=3, num_classes=num_classes, anchors=cfg.model.head.anchors).to(device)
+    return model
